Debug branch (#110)
* Test branch (#73)

* nothing

* repair Users and channel

* trying to repair kickuser

* repair erase for all

* resolve crash for MODE cmd

* add numeric replies for MODE cmd and change of testbranch

* numeric rpl mode

* finish join cmd, have to finish RPL on MODE cmd

* resolve conflit

* repair JOIN

* limit check

* rpl channel_mode is

* kick cmd

* kikick

* try to set JOIN_LIST

* kick cmd maj, approx

* joinjoin

* conflit

* kick message kicker

* more mode

* changes

* response MODE active

* changes

* add InviteList + new name of channel

* add delete channel with /quit and ft for Nickname check before connexion

* nice

* resolve kick with comment

* pointeur de mort

* okok

* /quit user from serverlist ope

+++ cmd/quit.cpp
@@ -6,10 +6,15 @@ bool	cmd::parseQuit(Server *server, std::string str, User * user)
 	if (arg.size() != 2) {
 		return false;
 	}
-	//delete all chan
-	server->deleteAllChannel();
-	std::string quit_message = ":" + user->getNickname() + "!" + user->getUsername() + "@localhost " + arg[0] + "\r\n";
-	send(user->getSocket(), quit_message.c_str(), quit_message.size(), 0);
+
+	if (server->getUserList().size() == 1) {
+		server->deleteAllChannel();
+		std::string quit_message = ":" + user->getNickname() + "!" + user->getUsername() + "@localhost " + arg[0] + "\r\n";
+		send(user->getSocket(), quit_message.c_str(), quit_message.size(), 0);
+	}
+
+	server->deleteUserFromUserList(*user);
+
 	std::cerr << "QUIT COMMAND" << std::endl;
 	return true;
 }
